requirements.txt
fastapi
uvicorn[standard]

requests
python-dotenv

parlant-sdk

pydantic
